Move FEC unit tests into their own folders
