+++ tests/codes/bch/test_decode.py
@@ -6,7 +6,7 @@ import numpy as np
 
 import galois
 
-from .helper import random_errors, random_type
+from ..helper import random_errors, random_type
 
 CODES = [
     (15, 11),  # GF(2^4) with t=1
